ai-driven-development/packages/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-dj-backend"
version = "0.1.0"
description = "Community backend for AI DJ Assist"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pyjwt[crypto]>=2.8",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "passlib[bcrypt]>=1.7",
    "slowapi>=0.1",
    "cryptography>=42.0",
    "stripe>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.10",
    "aiosqlite>=0.20",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]
ignore = ["B008"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
